backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.2
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
